Fix navigation paths and imports after merge
