+++ backend/data/visuals/elo_ratings_over_time.png
