+++ backend/data/visuals/glicko_ratings_over_time.png
