+++ backend/data/visuals/margin_home_elo_ratings_over_time.png
