+++ backend/data/visuals/trueskill_ratings_over_time.png
