+++ backend/rating_algorithms/__init__.py
@@ -1,0 +1,28 @@
+"""Rating algorithms package."""
+
+from .engines import (
+    RatingEngine,
+    GlickoEngine,
+    EloEngine,
+    MarginHomeElo,
+    TrueSkillEngine,
+)
+
+from .data_prep import (
+    load_joined_games,
+    explore_dataframe,
+    summarize_games,
+    build_results,
+)
+
+__all__ = [
+    "RatingEngine",
+    "GlickoEngine",
+    "EloEngine",
+    "MarginHomeElo",
+    "TrueSkillEngine",
+    "load_joined_games",
+    "explore_dataframe",
+    "summarize_games",
+    "build_results",
+]
