+++ frontend/src/App.jsx
@@ -1,15 +1,12 @@
-import { useState, useEffect } from 'react';
 import './index.css';
 import Layout from './components/Layout';
-import axios from 'axios';
-
 import { Routes, Route } from 'react-router-dom';
 import About from './pages/About';
 import Home from './pages/Home';
-import Contact from './pages/Contact'
-import DashboardHome from './pages/DashboardHome';
+import Contact from './pages/Contact';
 import GamePrediction from './pages/GamePrediction';
 import HistoricalRanking from './pages/HistoricalRanking';
+import TeamStats from './pages/TeamStats';
 
 function App() {
 
@@ -20,7 +17,7 @@ function App() {
         <Route path="/" element={<Home />} />
         <Route path="/about" element={<About />} />
         <Route path="/contact" element={<Contact />} />
-        <Route path="/dashboardhome" element={<DashboardHome />} />
+        <Route path="/teamstats" element={<TeamStats />} />
         <Route path="/gameprediction" element={<GamePrediction />} />
         <Route path="/historicalranking" element={<HistoricalRanking />} />
       </Routes>
